reorganise functions, pull out get-results from discord
