+++ commands/db/admin/set-submission-channel.py
@@ -4,7 +4,7 @@ import discord
 from discord.ext import commands
 from dotenv import load_dotenv
 
-from api.submissions import set_submission_channel
+from api.utils import set_submission_channel
 
 load_dotenv()
 DEFAULT = os.getenv('DEFAULT')  # Choices: mkw, sm64
