+++ commands/db/host/submit.py
@@ -5,8 +5,8 @@ from discord.ext import commands
 from dotenv import load_dotenv
 
 from api.submissions import post_submission_list, \
-    get_display_name, get_submission_channel, submit_file
-from api.utils import has_host_role
+    submit_file
+from api.utils import has_host_role, get_submission_channel, get_display_name
 from discord_ext.discord_utils import edit_submission_list
 
 load_dotenv()
